requirements.txt
Django>=2.0
psycopg2
django-leaflet
requests
django-after-response
